Consolidate user actions and user action tests into one file and update imports
