+++ public/src/containers/sub_apps/admin_container.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 
 import Admin from '../../components/sub_apps/admin';
 
-import { fetchUser } from '../../actions/account_info';
+import { fetchUser } from '../../actions/users';
 
 function mapStateToProps(state) {
   return {
