+++ public/src/containers/sub_apps/volunteers_container.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 
 import Volunteers from '../../components/sub_apps/volunteers';
 
-import { fetchUser } from '../../actions/account_info';
+import { fetchUser } from '../../actions/users';
 
 function mapStateToProps(state) {
   return {
